styles,pages and most of the conections done

+++ src/app/features/admin/components/tables/route-table/route-table.component.ts
@@ -1,4 +1,6 @@
 import { Component, Input, OnInit, ChangeDetectorRef, Output, EventEmitter } from '@angular/core';
+import { ConfirmDeleteAlertComponent } from '../../alerts/confirm-delete-alert/confirm-delete-alert.component';
+import { DeleteAlertComponent } from '../../alerts/delete-alert/delete-alert.component';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { RouteRepository } from '../../../data/repository/route-repository';
@@ -7,11 +9,14 @@ import { Route } from '../../../data/models/route.model';
 @Component({
   selector: 'app-route-table',
   standalone: true,
-  imports: [CommonModule, RouterModule],
+  imports: [CommonModule, RouterModule, ConfirmDeleteAlertComponent, DeleteAlertComponent],
   templateUrl: './route-table.component.html',
   styleUrls: ['./route-table.component.scss']
 })
 export class RouteTableComponent implements OnInit {
+  showConfirmDelete = false;
+  showDeleteAlert = false;
+  indexToDelete: number | null = null;
   @Output() asignarParadas = new EventEmitter<number>();
   @Input() token: string | undefined;
   rutas: Route[] = [];
@@ -47,20 +52,39 @@ export class RouteTableComponent implements OnInit {
     this.selectedIndex = index;
   }
 
-  eliminar(index: number) {
-    const id = this.rutas[index].id;
-    const tokenToUse = this.getToken();
-    this.repository.delete(id).subscribe({
-      next: () => {
-        this.loadRutas();
-        if (this.selectedIndex === index) {
-          this.selectedIndex = null;
+  onDeleteClick(index: number) {
+    this.indexToDelete = index;
+    this.showConfirmDelete = true;
+  }
+
+  onCancelDelete() {
+    this.showConfirmDelete = false;
+    this.indexToDelete = null;
+  }
+
+  onConfirmDelete() {
+    if (this.indexToDelete !== null) {
+      const id = this.rutas[this.indexToDelete].id;
+      this.repository.delete(id).subscribe({
+        next: () => {
+          this.loadRutas();
+          if (this.selectedIndex === this.indexToDelete) {
+            this.selectedIndex = null;
+          }
+          this.showConfirmDelete = false;
+          this.indexToDelete = null;
+          this.showDeleteAlert = true;
+        },
+        error: (err) => {
+          this.showConfirmDelete = false;
+          this.indexToDelete = null;
         }
-      },
-      error: (err) => {
-        console.error('[RouteTable] Error al eliminar ruta', err);
-      }
-    });
+      });
+    }
+  }
+
+  onCloseDeleteAlert() {
+    this.showDeleteAlert = false;
   }
 
   get selectedRoute() {
